Methods business rule adjustments

+++ api-rest/src/main/java/com/api/apirest/services/ApiRestService.java
@@ -23,6 +23,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
 import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.util.ArrayList;
@@ -307,6 +308,7 @@ public class ApiRestService {
         BeanUtils.copyProperties(totalDto, totalModel);
         totalModel.setCreatedDate(LocalDateTime.now(ZoneId.of("UTC")));
         totalModel.setExternalId(UUID.randomUUID().toString());
+        totalModel.setRemainder(totalModel.getTotal());
 
         //Adiciona a criança no novo total mensal que está criando
         childModel.addTotal(totalModel);
@@ -320,6 +322,7 @@ public class ApiRestService {
                 sponsorModel.getExternalId(),
                 childModel.getExternalId(),
                 totalModel.getTotal(),
+                totalModel.getRemainder(),
                 totalModel.getDescription()
         );
 
@@ -338,6 +341,7 @@ public class ApiRestService {
                         externaId,
                         childModelSponsor.get().getExternalId(),
                         totalModelSponsor.getTotal(),
+                        totalModelSponsor.getRemainder(),
                         totalModelSponsor.getDescription()
                 );
                 return ResponseEntity.status(HttpStatus.OK).body(respTotal);
@@ -354,6 +358,7 @@ public class ApiRestService {
                         externaId,
                         childModelSponsor.get().getExternalId(),
                         totalModelChild.getTotal(),
+                        totalModelChild.getRemainder(),
                         totalModelChild.getDescription()
                 );
                 return ResponseEntity.status(HttpStatus.OK).body(respTotal);
@@ -369,6 +374,11 @@ public class ApiRestService {
         TotalMonthlyAmountModel totalModel = totalRepository.findByExternalId(externalId);
         //Remover
         if(totalModel != null) {
+            if(taskRepository.existsByTotalModel(totalModel)) {
+                Messages messages = new Messages(messageProperty.getProperty("error.task.delete"), HttpStatus.BAD_REQUEST.value());
+
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(messages);
+            }
             totalModel.getSponsorModel().removeTotal(totalModel);
             totalModel.getChildModel().removeTotal(totalModel);
             totalRepository.delete(totalModel);
@@ -378,7 +388,7 @@ public class ApiRestService {
         }
         Messages messages = new Messages(messageProperty.getProperty("error.total.notFound"), HttpStatus.NOT_FOUND.value());
 
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(messages);
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(messages);
     }
 
 
@@ -396,23 +406,36 @@ public class ApiRestService {
                 bonusModel.setCreatedDate(LocalDateTime.now(ZoneId.of("UTC")));
                 bonusModel.setExternalId(UUID.randomUUID().toString());
 
+                //Atualiza o valor no total
+                BigDecimal getTotal = new BigDecimal(totaModel.getTotal());
+                BigDecimal bonusWeigth = new BigDecimal(bonusModel.getBonusWeigth());
+                BigDecimal divisor = new BigDecimal("100.0");
+                BigDecimal getRemainder = new BigDecimal(totaModel.getRemainder());
+                BigDecimal valueBonus = getTotal.multiply(bonusWeigth.divide(divisor));
+                BigDecimal total = getTotal.add(valueBonus);
+                total = total.setScale(2, RoundingMode.HALF_EVEN);
+                totaModel.setTotal(total.doubleValue());
+                valueBonus = valueBonus.setScale(2, RoundingMode.HALF_EVEN);
+
+                //Atualiza o valor do remainder
+                BigDecimal remainder = getRemainder.add(valueBonus);
+                remainder = remainder.setScale(2, RoundingMode.HALF_EVEN);
+                totaModel.setRemainder(remainder.doubleValue());
+
                 //Adiciona o responsável no novo bônus mensal que está criando
                 sponsorModel.addBonus(bonusModel);
-
                 //Salva o bônus no banco de dados
                 bonusRepository.save(bonusModel);
-
-                //Atualiza o valor no total
-                BigDecimal result = BigDecimal.valueOf(totaModel.getTotal()).add(BigDecimal.valueOf(bonusModel.getBonus()));
-                result = result.setScale(2, BigDecimal.ROUND_HALF_UP);
-                totaModel.setTotal(result.doubleValue());
+                //Salva o total atualizado no banco de dados
                 totalRepository.save(totaModel);
 
                 RespTotal respTotal = new RespTotal(
                         totaModel.getExternalId(),
                         sponsorModel.getExternalId(),
                         childModelSponsor.get().getExternalId(),
-                        result.doubleValue(),
+                        total.doubleValue(),
+                        valueBonus.doubleValue(),
+                        remainder.doubleValue(),
                         totaModel.getDescription()
                 );
 
@@ -420,11 +443,11 @@ public class ApiRestService {
             }
             Messages messages = new Messages(messageProperty.getProperty("error.total.notFound"), HttpStatus.NOT_FOUND.value());
 
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(messages);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(messages);
         }
         Messages messages = new Messages(messageProperty.getProperty("error.sponsor.notFound"), HttpStatus.NOT_FOUND.value());
 
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(messages);
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(messages);
     }
 
 
@@ -442,22 +465,42 @@ public class ApiRestService {
                 penaltyModel.setCreatedDate(LocalDateTime.now(ZoneId.of("UTC")));
                 penaltyModel.setExternalId(UUID.randomUUID().toString());
 
+                //Atualiza o valor no total
+                BigDecimal getTotal = new BigDecimal(totaModel.getTotal());
+                BigDecimal penaltyWeigth =  new BigDecimal(penaltyModel.getPenaltyWeigth());
+                BigDecimal divisor = new BigDecimal("100.0");
+                BigDecimal getRemainder = new BigDecimal(totaModel.getRemainder());
+                BigDecimal valuePenalty = getTotal.multiply(penaltyWeigth.divide(divisor));
+
+                if(valuePenalty.doubleValue() > getRemainder.doubleValue()) {
+                    Messages messages = new Messages(messageProperty.getProperty("error.penalty.addTask"), HttpStatus.BAD_REQUEST.value());
+                    return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(messages);
+                }
+
+                BigDecimal total = getTotal.subtract(valuePenalty);
+                total = total.setScale(2, RoundingMode.HALF_EVEN);
+                totaModel.setTotal(total.doubleValue());
+                valuePenalty = valuePenalty.setScale(2, RoundingMode.HALF_EVEN);
+
+                //Atualiza o valor do remainder
+                BigDecimal remainder = getRemainder.subtract(valuePenalty);
+                remainder = remainder.setScale(2, RoundingMode.HALF_EVEN);
+                totaModel.setRemainder(remainder.doubleValue());
+
                 //Adiciona o responsável na nova penalidade mensal que está criando
                 sponsorModel.addPenalty(penaltyModel);
-
                 //Salva a penalidade no banco de dados
                 penaltyRepository.save(penaltyModel);
-
-                //Atualiza o valor no total
-                double result = totaModel.getTotal() - penaltyModel.getPenalty();
-                totaModel.setTotal(result);
+                //Salva o total no banco de dados
                 totalRepository.save(totaModel);
 
                 RespTotal respTotal = new RespTotal(
                         totaModel.getExternalId(),
                         sponsorModel.getExternalId(),
                         childModelSponsor.get().getExternalId(),
-                        result,
+                        total.doubleValue(),
+                        valuePenalty.doubleValue(),
+                        remainder.doubleValue(),
                         totaModel.getDescription()
                 );
 
@@ -465,33 +508,90 @@ public class ApiRestService {
             }
             Messages messages = new Messages(messageProperty.getProperty("error.total.notFound"), HttpStatus.NOT_FOUND.value());
 
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(messages);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(messages);
         }
         Messages messages = new Messages(messageProperty.getProperty("error.sponsor.notFound"), HttpStatus.NOT_FOUND.value());
 
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(messages);
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(messages);
     }
 
 
     //Task
     @Transactional
-    public ResponseEntity<Object> createTask (TaskModel taskModel, String externalIdSponsor) {
-        try {
+    public ResponseEntity<Object> createTask (TaskDto taskDto) {
             //Busca o responsável com externalId passado
-            SponsorModel sponsorModel = sponsorRepository.findByExternalId(externalIdSponsor);
-            //Adiciona a nova tarefa ao Sponsor que está criando
-            sponsorModel.addTask(taskModel);
-            //Salva a criança no banco de dados
-            taskRepository.save(taskModel);
-            RespTask respTask = new RespTask(
-                    taskModel.getExternalId(), taskModel.getName(), taskModel.getDescription(), taskModel.getWeight()
-            );
+            SponsorModel sponsorModel = sponsorRepository.findByExternalId(taskDto.getExternalIdSponsor());
+            if(sponsorModel != null) {
+                //Busca a criança com externalId passado
+                ChildModel childModel = childRepository.findByExternalId(taskDto.getExternalIdChild());
+                if(childModel != null) {
+                    //Buasca a tarefa com externalId passado
+                    TotalMonthlyAmountModel totalModel = totalRepository.findByExternalId(taskDto.getExternalIdTotal());
+                    if(totalModel != null) {
+                        //Manipula os atributos do model
+                        TaskModel taskModel = new TaskModel();
+                        BeanUtils.copyProperties(taskDto, taskModel);
+                        taskModel.setCreatedDate(LocalDateTime.now(ZoneId.of("UTC")));
+                        taskModel.setExternalId(UUID.randomUUID().toString());
 
-            return ResponseEntity.status(HttpStatus.CREATED).body(respTask);
-        } catch (NullPointerException e) {
+                        //Fazer o cálculo para saber o restante em cima do total
+                        BigDecimal weight = new BigDecimal(taskModel.getWeight());
+                        BigDecimal divisor = new BigDecimal("100.0");
+                        BigDecimal total = new BigDecimal(totalModel.getTotal());
+                        BigDecimal valueTask = weight.divide(divisor, 2, RoundingMode.HALF_UP).multiply(total);
+
+                        //Se o peso for menor que o valor total
+                        if(valueTask.doubleValue() > totalModel.getRemainder()) {
+                            BigDecimal remainder = new BigDecimal(totalModel.getRemainder());
+
+                            BigDecimal weightRemainder = remainder.multiply(divisor).divide(total, 2, RoundingMode.HALF_UP);
+                            Messages messages = new Messages(messageProperty.getProperty("error.createTask.value") + weightRemainder, HttpStatus.BAD_REQUEST.value());
+
+                            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(messages);
+                        }
+                        BigDecimal remainder  = new BigDecimal(totalModel.getRemainder()).subtract(valueTask);
+                        remainder = remainder.setScale(2, RoundingMode.HALF_EVEN);
+                        totalModel.setRemainder(remainder.doubleValue());
+                        valueTask = valueTask.setScale(2, RoundingMode.HALF_EVEN);
+
+                        //Adiciona a nova tarefa ao Sponsor
+                        sponsorModel.addTask(taskModel);
+                        //Adiciona a nova tarefa a Child
+                        childModel.addTask(taskModel);
+                        //Adiciona a nova tarefa ao total
+                        totalModel.addTask(taskModel);
+                        //Edita o total com o valor do remainder atualizado
+                        totalRepository.save(totalModel);
+                        //Salva a tarefa no banco de dados
+                        taskRepository.save(taskModel);
+
+                        //Prepara a resposta
+                        RespTask respTask = new RespTask(
+                                taskModel.getExternalId(),
+                                taskModel.getSponsorModel().getExternalId(),
+                                taskModel.getChildModel().getExternalId(),
+                                taskModel.getTotalModel().getExternalId(),
+                                taskModel.getName(),
+                                taskModel.getDescription(),
+                                taskModel.getWeight(),
+                                valueTask.doubleValue(),
+                                taskModel.isComplete(),
+                                totalModel.getRemainder()
+                        );
+
+                        return ResponseEntity.status(HttpStatus.CREATED).body(respTask);
+                    }
+                    Messages messages = new Messages(messageProperty.getProperty("error.total.notFound"), HttpStatus.NOT_FOUND.value());
+
+                    return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(messages);
+                }
+                Messages messages = new Messages(messageProperty.getProperty("error.child.notFound"), HttpStatus.NOT_FOUND.value());
+
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(messages);
+            }
             Messages messages = new Messages(messageProperty.getProperty("error.sponsor.notFound"), HttpStatus.NOT_FOUND.value());
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(messages);
-        }
+
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(messages);
     }
